requirements.txt
celery==5.3.4
redis==5.0.1
docling==1.0.0
llama-index==0.9.0
llama-index-embeddings-openai==0.1.0
llama-index-vector-stores-chroma==0.1.0
requests==2.31.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic>=2.6.0,<3.0.0
loguru==0.7.2
aiofiles==23.2.1
httpx==0.25.2
# sqlite3 is built-in with Python
